refactor: move extract_file to common directory for reuse
- Extracted the `extract_file` function from apk.cpp
- Created separated header (extract_file.h) and source (extract_file.cpp) files in common directories
- Updated CMakeLists.txt files include the new files in the build

This change centralizes file extraction functionality for better reusability in the future implementation of IPA files.
