+++ shuriken/lib/common/CMakeLists.txt
@@ -1,4 +1,5 @@
 target_sources(shuriken-obj PRIVATE
     ${CMAKE_CURRENT_LIST_DIR}/shurikenstream.cpp
     ${CMAKE_CURRENT_LIST_DIR}/logger.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/extract_file.cpp
 )
